+++ FuzzyApprox.h
@@ -8,6 +8,8 @@
 #include "RuleList.h"
 #include "FunctionApprox.h"
 
+#include <time.h>
+
 #define ERR_MEMORY 20
 
 class CFuzzyFunctionApproximation : public CFunctionApproximation {
@@ -16,16 +18,16 @@ public:
   CFuzzyFunctionApproximation();
   ~CFuzzyFunctionApproximation();
 
-  bool Init(REAL _threshold, int min, int max, int nouts);
+  bool Init(double _threshold, int min, int max, int nouts);
   
   bool FuzzySetsInit(const vector<pair<int,int> >& min_max);
-  bool StartUp(int MaxNumberOfRules, REAL threshold, int _min, int _max);
+  bool StartUp(int MaxNumberOfRules, double threshold, int _min, int _max);
  
-  bool Learn(REAL* InputValues, REAL* OutputValues);
+  bool Learn(double* InputValues, double* OutputValues);
   bool Learn(Configuration conf,Simulation sim,Processor& p,Mem_hierarchy& mem);
   bool Learn(const Configuration& conf,const Dynamic_stats& dyn);
-  bool GenerateInputFuzzySets(int Dimensionality, int* InputFuzzySetsNumberVector, REAL *InMinimumValuesVector, REAL *InMaximumValuesVector);
-  bool EstimateG(REAL* InputValues,REAL* OutputVector);
+  bool GenerateInputFuzzySets(int Dimensionality, int* InputFuzzySetsNumberVector, double *InMinimumValuesVector, double *InMaximumValuesVector);
+  bool EstimateG(double* InputValues,double* OutputVector);
   Simulation Estimate1(Configuration conf,Processor& p,Mem_hierarchy& mem);
   Dynamic_stats Estimate2(Configuration conf);
   bool Reliable();
@@ -36,22 +38,22 @@ private:
   int position();
   int InDim,OutDim;
   int *InputSetsNumber;
-  REAL *InputsMin;
-  REAL *InputsMax;
-  REAL *InputCenters;
+  double *InputsMin;
+  double *InputsMax;
+  double *InputCenters;
   CRuleList *RuleTable;
   Rule newRule;
-  REAL *estimatedValues;
-  REAL *degrees;
-  REAL *alpha;
+  double *estimatedValues;
+  double *degrees;
+  double *alpha;
   int *Sets;
   int count;
   bool calcola;
-  REAL *errore;
-  REAL **errmatrix;
+  double *errore;
+  double **errmatrix;
   int prove;
-  REAL *stima;
-  REAL threshold;
+  double *stima;
+  double threshold;
   int min_sims,max_sims;
   int posx;
   FILE* fuzzy_log;
